I. #45 Add LULC legend control to map
Adding event handler to map for 'pointermove' that sets the state for a
LULC code the mouse pointer is hovering over. That code is then
displayed in a box on the bottom right of the map.

Added a new react component file to handle the display of the legend
control component.

The code style should only show when there is a LULC selected (visible)
and it should correspond to the LULC layer "in front". That is, if a
base LULC and Scenario LULC are both visible, the Scenario LULC should
be the one reported.